+++ src/web/assets/commerceui/src/js/order/apps/OrderNotices.vue
@@ -1,6 +1,35 @@
 <template>
-    <div v-if="showNotices">
-        <div class="meta read-only warning">
+    <div class="order-notices-wrapper">
+        <div v-if="showAdminNotices" class="meta read-only order-admin-notices">
+            <div class="order-flex order-notices-header">
+                <div>
+                    {{ 'Admin Notices' | t('commerce') }}
+                </div>
+            </div>
+            <div class="order-notices-items">
+                <div v-for="notice in adminNotices" :key="notice.id">
+                    <hr />
+                    <div
+                        class="order-flex order-notices-item order-admin-notice-item"
+                    >
+                        <div>{{ notice.message }}</div>
+                        <div>
+                            <button
+                                @click.prevent="dismissAdminNotice(notice)"
+                                class="btn small"
+                            >
+                                {{ 'Clear' | t('commerce') }}
+                            </button>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+        <div
+            v-if="showNotices"
+            class="meta read-only warning order-customer-notices"
+        >
             <div class="order-flex order-notices-header">
                 <div>
                     {{ 'Customer Notices' | t('commerce') }}
@@ -12,7 +41,7 @@
                 </div>
             </div>
             <div class="order-notices-items">
-                <div v-for="notice in draft.order.notices" :key="notice.id">
+                <div v-for="notice in regularNotices" :key="notice.id">
                     <hr />
                     <div class="order-notices-item">
                         {{ notice.message }}
@@ -38,17 +67,46 @@
                 },
             },
 
+            adminNotices() {
+                if (
+                    !this.draft ||
+                    !this.draft.order ||
+                    !this.draft.order.adminNotices
+                ) {
+                    return [];
+                }
+                return Object.values(this.draft.order.adminNotices);
+            },
+
+            regularNotices() {
+                if (
+                    !this.draft ||
+                    !this.draft.order ||
+                    !this.draft.order.notices
+                ) {
+                    return [];
+                }
+                return Object.values(this.draft.order.notices);
+            },
+
+            showAdminNotices() {
+                return this.adminNotices.length > 0;
+            },
+
             showNotices() {
-                return (
-                    this.draft &&
-                    this.draft.order &&
-                    this.draft.order.notices &&
-                    Object.keys(this.draft.order.notices).length
-                );
+                return this.regularNotices.length > 0;
             },
         },
 
         methods: {
+            dismissAdminNotice(notice) {
+                let draft = this.draft;
+                draft.order.adminNotices = this.adminNotices.filter(
+                    (n) => n.id !== notice.id
+                );
+                this.draft = draft;
+            },
+
             clearNotices() {
                 let draft = this.draft;
                 draft.order.notices = [];
@@ -73,5 +131,41 @@
 
     .order-notices-item {
         padding-bottom: 14px;
+    }
+
+    .order-notices-wrapper {
+        margin-bottom: var(--spacing);
+    }
+
+    .order-customer-notices {
+        box-shadow: 0 0 0 1px var(--yellow-300);
+    }
+
+    .order-admin-notices {
+        padding-block: var(--m);
+        color: var(--text-color) !important;
+        background-color: var(--red-050) !important;
+        box-shadow: 0 0 0 1px var(--red-300);
+        margin-bottom: 14px;
+
+        .btn {
+            background-color: var(--red-500);
+            color: #fff;
+
+            &:hover,
+            &:focus {
+                background-color: var(--red-600);
+            }
+
+            &:active {
+                background-color: var(--red-700);
+            }
+        }
+    }
+
+    .order-admin-notice-item {
+        align-items: center;
+        justify-content: space-between;
+        gap: 14px;
     }
 </style>